Move distribution methods; partitioned epoch-rewards reorg, 3 of 5 (#528)
* Add distribution sub-submodule

* Move distribution methods to sub-submodule

* Move unit tests into distribution sub-submodule
