+++ runtime/src/bank/partitioned_epoch_rewards/mod.rs
@@ -1,3 +1,4 @@
+mod distribution;
 mod sysvar;
 
 use {
